Update container env to be more explicit
Removes the `env_file` config from the compose file in order to make clearer the distinction between the compose file env vars and the container env vars. Instead of setting both to have the same .env file, this allows vars to be used only in the compose file, which then sets them as build args.