Add unique constraint for `Candidate` model
Adds unique constraints for the `Candidate` model that will allow for duplicate names as long as a birthdate is provided for at least one. This should probably be replaced with a unique ID of another type (rather than composite name-dob key) if the project grows to cover more regions, but should provide the right amount of flexibility and data consistency for the foreseeable future.