Add `Seat` model
Adds the `Seat` model and foreign keys on the `Candidate` model to represent the seat a candidate is running for, and the seat a candidate is currently holding, if any. The `Seat` model includes custom `clean` and `validate_unique` methods in order to validate provided data for different types of seats. This was preferable to creating multiple tables for different types of seats, and having database-level checks in place on each, in order to utilize foreign keys, such as in the case of the `Candidate` model. It also prevents significant overlap between different database tables with only slight differences in data and model behavior.